Update for the Radon Transformation
I tried to update the main.py with the radon transformation. On my laptop, my spyder was not updated, so it won't let me install the pip. So I tried it here. I am not sure if it's going to work or not. Can you let me know if it doesn't?
Changes:
from line 16-20
and 112-128